messari_tg_bot/requirements.txt
httpx>=0.25
python-telegram-bot>=21.0
python-dotenv>=1.0
python-docx>=1.0
pytest>=7.4
feedparser>=6.0
certifi>=2023.7.22
beautifulsoup4>=4.12.0
html2text>=2020.1.16
